refactor: extract @beautiful-tree/types pkg
