+++ @beautiful-tree/react/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@beautiful-tree/react",
-  "version": "0.1.3",
+  "version": "0.2.0",
   "private": false,
   "author": "Andres Correa Casablanca <castarco@coderspirit.xyz>",
   "license": "MIT",
@@ -37,9 +37,8 @@
     "dist"
   ],
   "scripts": {
-    "build": "rollup --config rollup.config.prod.mjs",
+    "build": "rm -rf dist && rollup --config rollup.config.prod.mjs",
     "build-storybook": "storybook build",
-    "install-githooks": "if [ -d .git ]; then git config core.hooksPath .hooks; fi",
     "lint": "tsc && eslint . --ext ts,mts,tsx --report-unused-disable-directives --max-warnings 0 && publint",
     "lint:eslint": "eslint . --ext ts,mts,tsx --report-unused-disable-directives --max-warnings 0",
     "lint:publint": "publint",
@@ -47,10 +46,12 @@
     "storybook": "storybook dev -p 6006",
     "test": "vitest run",
     "test:coverage": "vitest run --coverage",
-    "test:watch": "vitest"
+    "test:watch": "vitest",
+    "prepublishOnly": "pnpm run build"
   },
   "devDependencies": {
     "@arethetypeswrong/cli": "^0.10.2",
+    "@beautiful-tree/types": "workspace:0.2.0",
     "@coderspirit/eslint-config": "^1.2.1",
     "@rollup/plugin-terser": "^0.4.3",
     "@rollup/plugin-typescript": "^11.1.3",
