+++ @beautiful-tree/react/src/BeautifulTree.tsx
@@ -1,6 +1,6 @@
 import { edgesIterator, postOrderIterator } from './traversal'
 import { Fragment } from 'react'
-import type { Tree } from './types'
+import type { Tree } from '@beautiful-tree/types'
 import type { WrappedTreeWithLayout } from './layouts'
 import { computeSmartLayout } from './layouts'
 export { computeNaiveLayout, computeSmartLayout } from './layouts'
