+++ @beautiful-tree/types/.eslintrc.cjs
@@ -3,7 +3,6 @@
 module.exports = {
 	root: true,
 	env: {
-		browser: true,
 		es2020: true,
 	},
 	parser: '@typescript-eslint/parser',
@@ -13,21 +12,9 @@ module.exports = {
 		tsConfigRootDir: __dirname,
 		project: ['./tsconfig.json'],
 	},
-	plugins: [
-		'react-refresh',
-	],
 	extends: [
 		'@coderspirit',
-		'plugin:react/recommended',
-		'plugin:react/jsx-runtime',
-		'plugin:react-hooks/recommended',
-		'plugin:storybook/recommended',
 	],
-	settings: {
-		react: {
-			version: 'detect'
-		},
-	},
 	ignorePatterns: [
 		'.eslintrc.cjs',
 		'dist/**/*',
