+++ @beautiful-tree/types/.prettierrc.mjs
@@ -1,0 +1,3 @@
+import config from '@coderspirit/eslint-config/prettier'
+
+export default config
