+++ @beautiful-tree/types/package.json
@@ -1,0 +1,43 @@
+{
+  "name": "@beautiful-tree/types",
+  "version": "0.2.0",
+  "private": false,
+  "author": "Andres Correa Casablanca <castarco@coderspirit.xyz>",
+  "license": "MIT",
+  "types": "./dist/types.d.cts",
+  "exports": {
+    "import": {
+      "types": "./dist/types.d.mts"
+    },
+    "require": {
+      "types": "./dist/types.d.cts"
+    }
+  },
+  "files": [
+    "dist"
+  ],
+  "scripts": {
+    "build": "rm -rf dist && rollup --config rollup.config.prod.mjs",
+    "lint": "tsc && eslint . --ext ts,mts,tsx --report-unused-disable-directives --max-warnings 0 && publint",
+    "lint:eslint": "eslint . --ext ts,mts,tsx --report-unused-disable-directives --max-warnings 0",
+    "lint:publint": "publint",
+    "lint:tsc": "tsc",
+    "test": "true",
+    "test:coverage": "true",
+    "prepublishOnly": "pnpm run build"
+  },
+  "devDependencies": {
+    "@arethetypeswrong/cli": "^0.10.2",
+    "@coderspirit/eslint-config": "^1.2.1",
+    "@rollup/plugin-typescript": "^11.1.3",
+    "@types/node": "^20.6.1",
+    "@typescript-eslint/parser": "^6.7.0",
+    "eslint": "^8.49.0",
+    "prettier": "^3.0.3",
+    "publint": "^0.2.2",
+    "rollup": "^3.29.1",
+    "rollup-plugin-dts": "^5.3.1",
+    "tslib": "^2.6.2",
+    "typescript": "^5.2.2"
+  }
+}
