+++ @beautiful-tree/types/tsconfig.json
@@ -1,0 +1,6 @@
+{
+	"extends": "../../tsconfig.json",
+	"compilerOptions": {
+		"baseUrl": "./src",
+	},
+}
