+++ package.json
@@ -3,5 +3,8 @@
   "version": "0.1.3",
   "private": true,
   "author": "Andres Correa Casablanca <castarco@coderspirit.xyz>",
-  "license": "MIT"
+  "license": "MIT",
+  "scripts": {
+    "install-githooks": "if [ -d .git ]; then git config core.hooksPath .hooks; fi"
+  }
 }
