+++ tsconfig.json
@@ -1,0 +1,20 @@
+{
+	"extends": "@coderspirit/eslint-config/tsconfig",
+	"compilerOptions": {
+		"target": "es2020",
+		"lib": ["ES2020", "DOM", "DOM.Iterable"],
+
+		"module": "ES2020",
+		"moduleResolution": "Bundler",
+
+		"noEmit": true,
+	},
+	"exclude": [
+		".github/**/*",
+		".hooks/**/*",
+		"coverage/**/*",
+		"dist/**/*",
+		"node_modules/**/*",
+		"storybook-static/**/*",
+	]
+}
